16645 octets in length
Including ContentType (1 byte), ProtocolVersion (2 bytes), and length (2 bytes) headers a single TLS record may be up to 16645 octets in length.

The value in RFC 5216 is even more wrong, I tried to correct that. Ben is correct that the 256 expansion was not included. But I think the headers are 5 bytes, not 3. Need to confirm with Ben.