Update draft-ietf-emu-eap-tls13.xml
Adding more text on ticket lifetime as raised by Ben and discussed in the issue tracker. Added the text suggested by Mohit. Also added several sentences describing that a ticket contains an identity and a PSK lifetime. Added info a key share in the full authentication so it does not show up only in resumption. Added text that key share and psk mode affect the security properties. And added text that EAP-TLS 1.3 provides forward secrecy.