normative
I agree with Éric Vyncke's comment. The ietf-tls-oldversions-deprecate draft normatively updates RFC 5216. It should definitly be normative is possible. The only reason it was informative was that EAP-TLS 1.3 did not want or needed to wait for ietf-tls-oldversions-deprecate to be published. No that ietf-tls-oldversions-deprecate is ahead in the publication que, there is no reason to have it informal anymore.