EAP State Machine
Based on a number of comments from Bernard. This EAP state machine is related to Ben's DISCUSS and the EMU consensus call on alternate status indications and use of close_notify.

The update tries to inplement the suggestions from Bernard. The informal RFC 4137 is references but not normatively. Requirement in RFC 3478 to differentiate protected and unprotected result indicators is followed.

This pull request is related to the pull requests "Packet Modification Attacks" and "MUST Error alert"