Packet Modification Attacks
Bernard has provided links to academic paper describing attacks on availability and ask for updates to the draft describing this. The new paragraph do just that.

Added that most things can be spoofed was requested by several persons including Security AD Ben and several TLS people. This was misunderstandings. 

This is related to the pull request "EAP state machine"