Protected success indication application data 0x00
This was discussed and EMU WG IETF 110.

https://datatracker.ietf.org/meeting/110/materials/slides-110-emu-eap-tls-13-implementation-report-00
https://datatracker.ietf.org/meeting/110/materials/slides-110-emu-eap-tls-with-tls-13-00

Interop test was done by several implementors during the hackaton. The conclusions was the application data works well as a protected success indication and that close_notify is problematic. Alan DeKok stated that application data will not cause problems for other TLS based EAP methods. The consensus in the WG was to use application data 0x00 as a protected success indication on the EAP-TLS layer.

The application data is not controlling the TLS layer. After sending application data (or any other success indication) the EAP-TLS can only send EAP-SUCCESS.

The working group argued that the IESG and TLS WG resistance against use application data was as it was used in -13 when it promised that no more handshake messages was sent by the TLS layers. The EMU WG hopes and thinks that the IESG and the TLS WG will be ok with the current use of applicaiton data as a protected success indication on the EAP-TLS layer. 

Additional changes related to this (protected success and protected failure) will/is done in the EAP State Machnine PR.